pdf_knowledge_tool/requirements.txt
docling
chromadb>=0.5.0
tinydb>=4.8.0
sentence-transformers>=3.0.0
streamlit>=1.35.0
transformers>=4.41.0
torch>=2.3.0
pillow
einops
accelerate
